Backport changes; use new versioning scheme
backports updated build.yml, refactored Config, and the fix for #18.

new version scheme is `BIG.FIX.BUILD`

where BIG is incremented with the implementations of large suggestions or significant changes, FIX is increased with bugfixes since BIG, and BUILD is always set to the number of `build` workflows on the GitHub Actions page + 1.