Complete Util overhaul plus a lot of other stuff
split Util.java into a handful of smaller classes because they are more concise and clear.
also replaced the hacky and ugly usage of SignedMessage for lastMsg in ChatPatches.java to a new MessageData record for super easy access. this also helps backporting efforts that i will ultimately force myself to do later...
plus a TON of other stuff

closes #65, #68, and #66 for 1.19.3+!