Add Only Invasive Drafting option
This adds an option called "Only Invasive Drafting Toggle" which works in combination with the "Drafting Toggle" option. It implements a fix for an issue that bothers me. Often times on a minigame server when you type out a message to a friend and the server moves you you have to type it out again. This would be fixed by the "Message Drafting Toggle" option but it also adds another inconvenience that outweighs the benefits: You can no longer delete your typed message by pressing esc. The addition adds another option that allows you to have both, depending on which you prefer:
- always persistent messages
and
- persistent messages that can still be removed by pressing esc