Change duplicate counter implementation
This changes the implementation of the a`ddCounter()` method from an `@injection` in the head, to a `@modifyvariable` method. As the code already modifies the incoming text message variable in the `modifyMessage()` method. The primary objective of this change is to improve mod compatibility, as unexpected things happened with the old implementaiton where it would be recurvise (or not, I'm not sure?) and cause some issues.

The new method makes the `addcounter()` method a unique one that returns a the modified text with the changed message. it is then called in `modifyMessage()`, just before that method returns the modified message. This should lead to better compatibility and it is also simpler to follow the flow of the program.