Removed ChatLog#transformUUIDArrays
This was a method created in an uninformed context. After having done some digging, the method is entirely useless and in the (short) future should be entirely redundant. although the codec-ified chat log writes INT_ARRAY uuids, it is capable of reading both INT_ARRAY and STRING variants of uuids, which when ported to all versions shouldn't throw any exceptions, which is what the original method prevented.